test: move ecdhCurve coverage to its own file
node-tls-server.test.ts has a pre-existing failure on main (the
SNICallback-matches-bind-hostname case), which trips the pass-after
check when the new tests share that file. Move the ecdhCurve tests to
node-tls-ecdh-curve.test.ts alongside the existing per-option test
files (node-tls-no-cipher-match-error.test.ts etc.).
